requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
agent-framework-core==1.0.0b260107
agent-framework-azure-ai==1.0.0b260107
